refactor(logic): split order_suggestion_move_army into part-of concern files (#3290)
Phase 0 prerequisite for the colonizethis_logic package split epic:
decompose the 538-line order_suggestion_move_army.dart (>500 lines) into
part-of concern files, following the established fog_resolution /
order_suggestion_work pattern. Pure structural extraction; no behavior change.

- order_suggestion_move_army.dart: thin library file (imports + part
  directives + concern overview doc).
- order_suggestion_move_unit.dart: civilian unit move suggestions
  (sortedMoveDestinationCandidateTileKeys, suggestMoveOrders).
- order_suggestion_army_move.dart: army move suggestions + Move Army picker
  (armyMoveCandidateDestinationProvinceIds, ArmyMovePickerDestination,
  armyMovePickerDestinations, suggestArmyMoveOrders).

All public symbols remain importable from the same library, so the
`export ... show` in order_suggestion.dart is unchanged.

Refs #3290
